CRUD operations added for Sports, Student, Coach tables

+++ src/objects/org/Player.java
@@ -14,8 +14,7 @@ public class Player {
     private int coachID , sportID, studentID;
     private String playerID;
     private Connection con = null;
-    private PreparedStatement statement = null;
-    private ResultSet resultSet = null;
+
     List<Player> players = new ArrayList<>();
 
     public Player(){}
@@ -58,7 +57,7 @@ public class Player {
     public void setStudentID(int studentID) {
         this.studentID = studentID;
     }
-
+    Student student = new Student();
     public void showStudentApplicant(){
         Student student = new Student();
         try {
@@ -74,43 +73,102 @@ public class Player {
         }
     }
 
-    public String generatePlayerID(){
-         int ID = countPlayers() + 1;
-        return "PL" + String.format("%04d",ID);
+    public boolean playerExists(int studentID){
+        PreparedStatement statement = null;
+        ResultSet resultSet = null;
+        String query = "SELECT * FROM PLAYERLIST";
+        try {
+            con = SetConnection.getConnection();
+            statement = con.prepareStatement(query);
+            resultSet = statement.executeQuery();
+            while (resultSet.next()) {
+                int s = Integer.parseInt(resultSet.getString("studentID"));
+                if (s == studentID)
+                    return true;
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+        return false;
     }
-    public int countPlayers(){
-        int counter= 0;
-        String query = "SELECT COUNT(playerID) FROM PLAYERLIST";
+    public boolean playerExists(String playerID){
+        PreparedStatement statement = null;
+        ResultSet resultSet = null;
+        String query = "SELECT * FROM PLAYERLIST";
+        try {
+            con = SetConnection.getConnection();
+            statement = con.prepareStatement(query);
+            resultSet = statement.executeQuery();
+            while (resultSet.next()) {
+                String s = resultSet.getString("playerID");
+                if (s.equalsIgnoreCase(playerID))
+                    return true;
+            }
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+        return false;
+
+    }
+    public String getLastPlayerID(){
+        PreparedStatement statement = null;
+        ResultSet resultSet = null;
+        String lastValue;
+        String query = "SELECT PLAYERID FROM PLAYERLIST ORDER BY PLAYERID DESC LIMIT 1";
         try {
             con = SetConnection.getConnection();
             statement = con.prepareStatement(query);
             resultSet = statement.executeQuery();
             resultSet.next();
-            counter = resultSet.getInt(1);
+            lastValue = resultSet.getString("playerID");
 
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
-        return counter;
+        return lastValue;
+
     }
 
-    //ON-HOLD
-    public void acceptPlayers(String studentID){
-        String pID = generatePlayerID();
-        String query = "INSERT INTO `playerlist` (`playerID`, `studentID`, `coachID`, `sportID`) VALUES (?, ?, ?, ?);";
-        try {
-            con = SetConnection.getConnection();
-            statement = con.prepareStatement(query);
-            statement.setString(1, pID);
-            statement.setString(2, studentID);
-            resultSet = statement.executeQuery();
+    public String generatePlayerID(){
+        String lastPlayer = getLastPlayerID();
+        StringBuffer number = new StringBuffer();
+        for (int i = 0 ; i < lastPlayer.length(); i ++ ){
+            if(Character.isDigit(lastPlayer.charAt(i)))
+                number.append(lastPlayer.charAt(i));
+        };
+        return "PL" + String.format("%04d",Integer.parseInt(String.valueOf(number)) + 1);
+    }
 
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+
+    public void acceptPlayers(Player player){
+        if(playerExists(player.getStudentID()) == true){
+            System.out.println("Player is already accepted.");
+        }
+        else if (student.studentExist(player.getStudentID()) == false){
+            System.out.println("There is no applicant student with that given ID number.");
+        }
+        else{
+            PreparedStatement statement = null;
+            ResultSet resultSet = null;
+            String query = "INSERT INTO `playerlist` (`playerID`, `studentID`, `coachID`, `sportID`) VALUES (?, ?, ?, ?);";
+            try {
+                con = SetConnection.getConnection();
+                statement = con.prepareStatement(query);
+                statement.setString(1, player.getPlayerID());
+                statement.setString(2, String.valueOf(player.getStudentID()));
+                statement.setString(3, String.valueOf(player.getCoachID()));
+                statement.setString(4, String.valueOf(player.getSportID()));
+                statement.execute();
+                System.out.println("Player " + player.getPlayerID()  + " is successfully added to the database.");
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
         }
     }
 
     public String showPlayerList() {
+        PreparedStatement statement = null;
+        ResultSet resultSet = null;
         String query = "SELECT * FROM PLAYERLIST";
         try {
             con = SetConnection.getConnection();
@@ -120,6 +178,7 @@ public class Player {
                 Player p = new Player(resultSet.getString("playerID"), Integer.valueOf(resultSet.getString("studentID")), Integer.valueOf(resultSet.getString("coachID")), Integer.valueOf(resultSet.getString("sportID")));
                 players.add(p);
             }
+            System.out.println("\t\tLIST OF PLAYERS:");
             System.out.printf("%-15s%-15s%-15s%-15s%n", "PlayerID", "StudentID", "CoachID", "SportID");
             players.forEach((p) -> System.out.print(p));
 
@@ -129,17 +188,23 @@ public class Player {
         return "";
     }
     public void removePlayer(String playerID){
-        String query = "DELETE FROM PLAYERLIST WHERE PLAYERID = ?";
-        try {
-            con = SetConnection.getConnection();
-            statement = con.prepareStatement(query);
-            statement.setString(1, playerID);
-            resultSet = statement.executeQuery();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+        if(playerExists(playerID) == false){
+            System.out.println("There is no player " + playerID + " in the database;" );
         }
-        System.out.println();
-
+        else {
+            PreparedStatement statement = null;
+            ResultSet resultSet = null;
+            String query = "DELETE FROM PLAYERLIST WHERE PLAYERID = ?";
+            try {
+                con = SetConnection.getConnection();
+                statement = con.prepareStatement(query);
+                statement.setString(1, playerID);
+                statement.execute();
+                System.out.println("A player is successfully deleted from database.");
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
+        }
     }
 
     @Override
@@ -149,8 +214,18 @@ public class Player {
 
     public static void main(String[] args) {
         Player p = new Player();
+        /*String ID = p.generatePlayerID();
+        p.setPlayerID(ID);
+        p.setCoachID(1101013);
+        p.setStudentID(2200465);
+        p.setSportID(1);
+        p.acceptPlayers(p);*/
+        //p.removePlayer("PL0041");
+        //System.out.println( p.playerExists(2200465));
+
         //System.out.println(p.generatePlayerID());
         //System.out.println(p.countPlayers());
-        System.out.println(p.showPlayerList());
+        //System.out.println(p.showPlayerList());
+        System.out.println(p.generatePlayerID());
     }
 }
